# docs/openapi/user_api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "User Services API",
    "description": "API for user management, authentication, collections, currency, and market.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:5001",
      "description": "Local API Gateway"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/user/{user_id}/market_list": {
      "get": {
        "summary": "Get the list of open markets",
        "tags": [
          "Market"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of open markets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "market_id": {
                        "type": "integer"
                      },
                      "item_id": {
                        "type": "integer"
                      },
                      "bid": {
                        "type": "number"
                      },
                      "seller_user_id": {
                        "type": "integer"
                      },
                      "status": {
                        "type": "string"
                      },
                      "end_date": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/{user_id}/transactions_history": {
      "get": {
        "summary": "Retrieve the transaction history of a user",
        "tags": [
          "Market"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "transaction_id": {
                        "type": "integer"
                      },
                      "item_id": {
                        "type": "integer"
                      },
                      "seller_user_id": {
                        "type": "integer"
                      },
                      "buyer_user_id": {
                        "type": "integer"
                      },
                      "status": {
                        "type": "string"
                      },
                      "start_date": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "end_date": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "bid": {
                        "type": "number"
                      },
                      "final_price": {
                        "type": "number"
                      },
                      "instance_id": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/{user_id}/instance/{instance_id}/auction": {
      "put": {
        "summary": "Create a new auction for an item instance",
        "tags": [
          "Market"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "description": "ID of the user creating the auction",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "description": "ID of the item instance to be auctioned",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Data required to create the auction",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "end_date": {
                    "type": "string",
                    "description": "Auction end date in `dd/mm/YYYY HH:MM` format"
                  },
                  "start_bid": {
                    "type": "number",
                    "description": "Initial bid amount, must be a positive number"
                  }
                },
                "required": [
                  "end_date",
                  "start_bid"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Auction created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "auction_id": {
                      "type": "integer"
                    },
                    "instance_id": {
                      "type": "integer"
                    },
                    "seller_user_id": {
                      "type": "integer"
                    },
                    "start_date": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "end_date": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "status": {
                      "type": "string"
                    },
                    "bid": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid input or auction already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "end_date and start_bid are mandatory"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized or invalid item instance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unauthorized or invalid instance"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/{user_id}/market/{market_id}/bid": {
      "put": {
        "summary": "Place a bid on an item in the market",
        "tags": [
          "Market"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "description": "ID of the user placing the bid",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "market_id",
            "in": "path",
            "required": true,
            "description": "ID of the market item",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Bid data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bid_amount": {
                    "type": "number"
                  }
                },
                "required": [
                  "bid_amount"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bid placed successfully"
          },
          "400": {
            "description": "Invalid bid or auction expired"
          },
          "403": {
            "description": "Cannot bid on own auction"
          },
          "404": {
            "description": "Auction not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}